+++ .github/workflows/CI.yml
@@ -36,8 +36,11 @@ jobs:
       - uses: actions/checkout@v3
       - name: Building
         run: |
-          cmake -DBUILD_TESTS=ON -B build
-          cmake --build build
+          mkdir build
+          d build/
+          cmake .. -DTEST=1
+          make --build .
+          cd ../
       
       - name: Upload our build dir
         uses: actions/upload-artifact@v3
