+++ CMakeLists.txt
@@ -1,15 +1,37 @@
-cmake_minimum_required(VERSION 3.14)
+cmake_minimum_required(VERSION 3.15)
 
-project(mutual_editing)
+project(chat_project)
 
-option(BUILD_TESTS "Build test programs" ON)
+set(CMAXE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+if(TEST)
+    set(CMAKE_CXX_FLAGS ${CMAKE_C_FLAGS} "-Wall -Wpedantic -Wextra -Werror")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fPIC -O0 -g")
+    set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -coverage -lgcov")
+    # set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -coverage -lgcov" )
+elseif()
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fPIC -O0")
+endif()
 
-set(CMAKE_CXX_FLAGS ${CMAKE_CXX_FLAGS} "-Wall -Wextra -Wpedantic -Werror")
 
-add_subdirectory("client_side")
-add_subdirectory("server_side")
+# set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -O0")
 
-if(BUILD_TESTS)
+add_subdirectory(Chat)
+
+add_executable(${PROJECT_NAME} ./source/main.cpp)
+
+
+
+target_include_directories(${PROJECT_NAME} PUBLIC ${MATRIX_VEC_LIB_INCLUDE_DIRS})
+target_link_libraries(${PROJECT_NAME} PUBLIC ${MATRIX_VEC_LIB_LIBRARIES})
+
+
+
+if(TEST)
+    message("-----MAIN-----")
+    message("Lib link = " ${MATRIX_VEC_LIB_LIBRARIES})
+    message("Dir link = " ${MATRIX_VEC_LIB_INCLUDE_DIRS})
+    message("---------------")
     enable_testing()
-    add_subdirectory(tests)
-endif(BUILD_TESTS)
+    add_subdirectory(test)
+endif()
