--- README.md
@@ -1,6 +1,0 @@
-# 2022_2_24-7
-# github-actions-tutorial
-[![CMake](https://github.com/cpp-park-vk-education/2022_2_24-7/actions/workflows/cmake.yml/badge.svg)](https://github.com/cpp-park-vk-education/2022_2_24-7/actions/workflows/cmake.yml)
-TeamWork - Сервис (ide) для совместного редактирования кода с поддержкой чата
-
-UML diagramm: https://app.diagrams.net/#G19ZA9EjK7nmwLxfxvh30ptExsg6YqAtec
