+++ Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Nov 21 05:41 MSK
+Start testing: Oct 24 18:01 MSK
 ----------------------------------------------------------
-End testing: Nov 21 05:41 MSK
+End testing: Oct 24 18:01 MSK
