--- client_side/CMakeLists.txt
@@ -1,8 +1,0 @@
-cmake_minimum_required(VERSION 3.14)
-
-project(client_side)
-
-set(CMAKE_CXX_STANDARD 17)
-
-set(SOURCES ./src/main.cpp)
-set(HEADERS ./include/main.hpp)
