--- client_side/src/main.cpp
@@ -1,3 +1,0 @@
-int main() {
-  return 0;
-}
