+++ test/test.cpp
@@ -1,0 +1,6 @@
+#include <gtest/gtest.h>
+#include "chat.hpp"
+
+TEST(ALG_Task1, count1) {
+    EXPECT_EQ((2*2), 4);
+}
