--- tests/testExampleLib/tests.cpp
@@ -1,9 +1,0 @@
-#include <gtest/gtest.h>
-
-TEST(DefaultTest, Test) {
-    EXPECT_EQ((2 * 2), 4);
-}
-
-TEST(DefaultTest2, Test) {
-    EXPECT_EQ((2 * 2), 4);
-}
